// @types/generated/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from "contentful";
import { Document } from "@contentful/rich-text-types";

export interface IPetFields {
  /** Title */
  title: string;

  /** Slug */
  slug: string;

  /** Thumbnail */
  thumbnail: Asset;

  /** Featured Image */
  featuredImage: Asset;

  /** Characteristics */
  characteristics: string[];

  /** Age */
  age: number;

  /** Bio */
  bio: Document;
}

export interface IPet extends Entry<IPetFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "pet";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export type CONTENT_TYPE = "pet";

export type LOCALE_CODE = "en-US";

export type CONTENTFUL_DEFAULT_LOCALE_CODE = "en-US";